refactor(journals-router): rename 'journal-route' to 'journals-router'
